pretty printing implemented
There are now '`formatted`' properties for metadatasets, metavars and distributions. These return information in an easy to read format.

There is also a function for the metadataset, "`preview_output()`" which prints the (to be generated) json to the console in a pretty/indented format (similar to what prettyprint would produce).